refactor(auth): improve code organization at auth service
