+++ src/api/auth/model/login-request.model.ts
@@ -1,0 +1,4 @@
+export interface LoginRequest {
+    email: string;
+    password: string;
+}
