+++ src/api/auth/model/login-response.model.ts
@@ -1,0 +1,4 @@
+export interface LoginResponse {
+    token: string;
+    refresh_token: string;
+}
