+++ src/api/user/model/create.model.ts
@@ -1,0 +1,5 @@
+export interface Create {
+    name: string;
+    email: string;
+    password: string;
+}
